--- a/src/WrapRootElement.tsx
+++ b/src/WrapRootElement.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import './styles/panda-prismjs.css';
+import { IntlContextProvider } from './context/react-intl';
+
+export const WrapRootElement: React.FC = ({ children }) => {
+  return <IntlContextProvider>{children}</IntlContextProvider>;
+};