--- a/src/WrapRootElement.tsx
+++ b/src/WrapRootElement.tsx
@@ -2,7 +2,12 @@ import React from 'react';
 
 import './styles/panda-prismjs.css';
 import { IntlContextProvider } from './context/react-intl';
+import { ThemeProvider } from './config/theme';
 
 export const WrapRootElement: React.FC = ({ children }) => {
-  return <IntlContextProvider>{children}</IntlContextProvider>;
+  return (
+    <ThemeProvider>
+      <IntlContextProvider>{children}</IntlContextProvider>
+    </ThemeProvider>
+  );
 };