--- a/src/components/AuthorPresentation/AuthorPresentation.tsx
+++ b/src/components/AuthorPresentation/AuthorPresentation.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { FormattedMessage } from 'gatsby-plugin-intl';
+import { FormattedMessage } from 'react-intl';
 import { Twitter, Linkedin, Github } from '@styled-icons/boxicons-logos';
 
 import * as S from './styled';