… .github/workflows/ci.yml
… gatsby-ssr.js
… helpers/createFields.test.js
… helpers/series.test.js
… package.json
… src/pages/index.tsx
… src/templates/blog-post.tsx
… src/types/GraphQL.ts
… src/types/Locales.ts
… src/types/Post.ts
… src/types/PostQuery.ts
… src/utils/index.ts
… tsconfig.json
… yarn.lock