gatsby-plugin-intl is an overkill for dynamic posts.

It generates pages for different locales by default, which means even if I don't have a translation, it creates a page for the article language and all other  supported.

If I have written 10 articles, 20 pages will be created, 10 for `pt-br` and 10 for `en`, even tho I don't have write any `en` page yet.

It also creates for all other pages like `/pages/`, `404`, etc. 

What I want is having an object with locales and allow the user switch between pt and en. When user click in a post, a information saying "there's a translation for this post" will be displayed within its link.