@raulfdm 2020-03-07
Still missing:

- [ ] when it has 2 versions (pt and en), it shows twice on Index and Search;
- [ ] first language is set to be `en` and does not consider user language