@raulfdm 2020-03-07
| Log                                    | Before                    | After                     | Result |
| -------------------------------------- | ------------------------- | ------------------------- |-------|
| success run queries                    | 19.944s - 87/87 4.36/s    | 21.477s - 29/29 1.35/s    | -58 pages/queries
| success Building static HTML for pages | 4.020s - 87/87 21.64/s    | 3.380s - 29/29 8.58/s     | -58 pages built
| success onPostBuild                    | 5.277s                    | 0.032s                    | -5.245s
| info Done building in                  | 140.5552074 sec           | 136.3574058 sec           | -4.1978s total (3% faster)

Results aren't that great but avoid creating useless pages and removing code to render or not page content (to avoid errors for pages without translation) its the goal.